abstracted helper methods into a setup class. Adjusted unused imports.
