+++ setup/reseed.js
@@ -1,6 +1,5 @@
-import Connection from "../api/db/connection.js";
 import databases_seed from "./seed/databases.js";
-import mongoose from "mongoose";
+import SetupHelper from "./helpers.js";
 
 import RegisteredUserSchema from "../api/models/user.registered.js";
 import UnregisteredUserSchema from "../api/models/user.unregistered.js";
@@ -8,7 +7,7 @@ import UnregisteredUserSchema from "../api/models/user.unregistered.js";
 import registered_users_seed from "./seed/registered_users_seed.js";
 import unregistered_users_seed from "./seed/unregistered_users_seed.js";
 
-const allConnections = openConnections(databases_seed);
+const allConnections = SetupHelper.openConnections(databases_seed);
 
 createDocuments(RegisteredUserSchema, registered_users_seed);
 createDocuments(UnregisteredUserSchema, unregistered_users_seed);
@@ -17,21 +16,7 @@ setTimeout(() => {
   process.exit();
 }, 2000);
 
-//closeConnections(allConnections);
-
-/**
- *
- * @param {*} databases
- * @returns
- */
-function openConnections(databases) {
-  let connections = [];
-  databases.forEach((db) => {
-    let connection = Connection.open(db);
-    connections.push(connection);
-  });
-  return connections;
-}
+//SetupHelper.closeConnections(allConnections);
 
 /**
  *
@@ -41,15 +26,5 @@ function openConnections(databases) {
 function createDocuments(schema, documents) {
   documents.forEach((doc) => {
     schema.create(doc);
-  });
-}
-
-/**
- *
- * @param {*} connections
- */
-function closeConnections(connections) {
-  connections.forEach((connection) => {
-    Connection.close(connection);
   });
 }
